Move pageable, followable interfaces to common package
