+++ models/src/main/java/com/vimeo/networking2/Category.kt
@@ -1,5 +1,6 @@
 package com.vimeo.networking2
 
+import com.vimeo.networking2.common.Followable
 import java.util.*
 
 /**
