+++ models/src/main/java/com/vimeo/networking2/CinemaList.kt
@@ -1,5 +1,7 @@
 package com.vimeo.networking2
 
+import com.vimeo.networking2.common.Pageable
+
 /**
  * List of cinema that could be paged.
  */
