+++ models/src/main/java/com/vimeo/networking2/ProductList.kt
@@ -1,5 +1,7 @@
 package com.vimeo.networking2
 
+import com.vimeo.networking2.common.Pageable
+
 /**
  * List of products that could be paged.
  */
